+++ tests/core/scene/a-scene.test.js
@@ -380,12 +380,16 @@ suite('a-scene (without renderer)', function () {
     });
   });
 
-  suite('reload', function () {
+  suite.only('reload', function () {
     test('reload scene innerHTML to original value', function () {
+      var canvasEl;
       var sceneEl = this.el;
       sceneEl.innerHTML = 'NEW';
       sceneEl.reload();
-      assert.equal(sceneEl.innerHTML, '');
+      assert.equal(sceneEl.children.length, 1);
+      canvasEl = sceneEl.querySelector('canvas');
+      assert.equal(canvasEl.getAttribute('class'), 'a-canvas');
+      assert.equal(canvasEl.getAttribute('data-aframe-canvas'), 'true');
     });
 
     test('reloads the scene and pauses', function () {
@@ -602,6 +606,16 @@ suite('scenes', function () {
       });
     });
     document.body.appendChild(sceneEl);
+  });
+});
+
+suite('setupCanvas', function () {
+  test('adds canvas to a-scene element', function () {
+    var el = this.sceneEl = document.createElement('a-scene');
+    el.canvas = undefined;
+    assert.notOk(el.canvas);
+    el.setupCanvas();
+    assert.ok(el.canvas);
   });
 });
